refactor(log): move SpyTransport and assertMessage into a separate file
