+++ tsconfig.test.json
@@ -1,7 +1,8 @@
 {
 	"extends": "./tsconfig.json",
 	"compilerOptions": {
-		"noImplicitAny": false
+		"noImplicitAny": false,
+		"types": ["jest-extended"]
 	},
 	"include": ["./**/*.test.ts"]
 }
